fix: image selector hanging
